+++ rechu/command/__init__.py
@@ -4,6 +4,7 @@ Subcommand collection package.
 
 from .base import Base
 from .create import Create
+from .new import New
 from .read import Read
 
 __all__ = ["Base"]