+++ tests/io/base.py
@@ -3,9 +3,11 @@ Tests for abstract base classes for file reading, writing and parsing.
 """
 
 from io import StringIO
+import os
 from pathlib import Path
 import unittest
-from rechu.io.base import Reader
+from rechu.io.base import Reader, Writer
+from rechu.models import Base
 
 class ReaderTest(unittest.TestCase):
     """
@@ -27,3 +29,24 @@ class ReaderTest(unittest.TestCase):
 
         with self.assertRaises(NotImplementedError):
             Reader(Path('fake/file')).parse(StringIO(''))
+
+class WriterTest(unittest.TestCase):
+    """
+    Tests for file writer.
+    """
+
+    def test_write(self) -> None:
+        """
+        Test writing the model to the path.
+        """
+
+        with self.assertRaises(NotImplementedError):
+            Writer(Path(os.devnull), Base()).write()
+
+    def test_serialize(self) -> None:
+        """
+        Test writing a serialized variant of a model to an open file.
+        """
+
+        with self.assertRaises(NotImplementedError):
+            Writer(Path('fake/file'), Base()).serialize(StringIO(''))